Struct primal_bit::BitVec

pub struct BitVec { /* fields omitted */ }

The bitvector type.

Implementations

impl BitVec

pub fn new() -> BitVec

Creates an empty BitVec.

Examples

use std::collections::BitVec;
let mut bv = BitVec::new();

pub fn from_u64s(data: Vec<u64>, bits: usize) -> BitVec

pub fn as_bytes_mut(&mut self) -> &mut [u8]

pub fn as_bytes(&self) -> &[u8]

pub fn as_u64s(&self) -> &[u64]

pub fn count_ones_before(&self, bit: usize) -> usize

Count the number of ones for the bits up to but not including the bitth bit.

pub fn find_nth_bit(&self, n: usize) -> Option<usize>

Find the index of the nth (0-indexed) set bit.

pub fn from_elem(nbits: usize, bit: bool) -> BitVec

Creates a BitVec that holds nbits elements, setting each element to bit.

Examples

use std::collections::BitVec;

let mut bv = BitVec::from_elem(10, false);
assert_eq!(bv.len(), 10);
for x in bv.iter() {
    assert_eq!(x, false);
}

pub fn get(&self, i: usize) -> Option<bool>

Retrieves the value at index i, or None if the index is out of bounds.

Examples

use std::collections::BitVec;

let bv = BitVec::from_bytes(&[0b01100000]);
assert_eq!(bv.get(0), Some(false));
assert_eq!(bv.get(1), Some(true));
assert_eq!(bv.get(100), None);

// Can also use array indexing
assert_eq!(bv[1], true);

pub fn set(&mut self, i: usize, x: bool)

Sets the value of a bit at an index i.

Panics

Panics if i is out of bounds.

Examples

use std::collections::BitVec;

let mut bv = BitVec::from_elem(5, false);
bv.set(3, true);
assert_eq!(bv[3], true);

pub unsafe fn set_unchecked(&mut self, i: usize, x: bool)

pub fn set_all(&mut self)

Sets all bits to 1.

Examples

use std::collections::BitVec;

let before = 0b01100000;
let after  = 0b11111111;

let mut bv = BitVec::from_bytes(&[before]);
bv.set_all();
assert_eq!(bv, BitVec::from_bytes(&[after]));

pub fn iter(&self) -> Iter<'_>

Returns an iterator over the elements of the vector in order.

Examples

use std::collections::BitVec;

let bv = BitVec::from_bytes(&[0b01110100, 0b10010010]);
assert_eq!(bv.iter().filter(|x| *x).count(), 7);

pub fn len(&self) -> usize

Returns the total number of bits in this vector

pub fn is_empty(&self) -> bool

Returns true if there are no bits in this vector

pub fn clear(&mut self)

Clears all bits in this vector.